add select authtype screen and update components and tests
