+++ app/src/organisms/SetupNetwork/ConnectedNetworkInfo.tsx
