+++ app/src/organisms/SetupNetwork/__tests__/ConnectedNetworkInfo.test.tsx
