+++ app/src/organisms/SetupNetwork/__tests__/SelectAuthenticationType.test.tsx
@@ -1,15 +1,17 @@
 import * as React from 'react'
-import { MemoryRouter, Route } from 'react-router-dom'
+import { MemoryRouter } from 'react-router-dom'
 import { fireEvent } from '@testing-library/react'
 
 import { renderWithProviders } from '@opentrons/components'
 
 import { i18n } from '../../../i18n'
 import * as Networking from '../../../redux/networking'
+import { SetWifiCred } from '../SetWifiCred'
 import { SelectAuthenticationType } from '../SelectAuthenticationType'
 
 const mockPush = jest.fn()
 
+jest.mock('../../../organisms/SetupNetwork/SetWifiCred')
 jest.mock('../../../redux/networking')
 jest.mock('../../../redux/discovery/selectors')
 jest.mock('react-router-dom', () => {
@@ -30,13 +32,14 @@ const initialMockWifi = {
 const mockGetNetworkInterfaces = Networking.getNetworkInterfaces as jest.MockedFunction<
   typeof Networking.getNetworkInterfaces
 >
+const mockSetWifiCred = SetWifiCred as jest.MockedFunction<typeof SetWifiCred>
 
-const render = (path = '/') => {
+const render = (
+  props: React.ComponentProps<typeof SelectAuthenticationType>
+) => {
   return renderWithProviders(
-    <MemoryRouter initialEntries={[path]} initialIndex={0}>
-      <Route path="/network-setup/wifi/select-auth-type/:ssid">
-        <SelectAuthenticationType />
-      </Route>
+    <MemoryRouter>
+      <SelectAuthenticationType {...props} />
     </MemoryRouter>,
     {
       i18nInstance: i18n,
@@ -45,18 +48,22 @@ const render = (path = '/') => {
 }
 
 describe('SelectAuthenticationType', () => {
+  let props: React.ComponentProps<typeof SelectAuthenticationType>
   beforeEach(() => {
+    props = {
+      ssid: 'mockWifi',
+      fromWifiList: undefined,
+    }
     mockGetNetworkInterfaces.mockReturnValue({
       wifi: initialMockWifi,
       ethernet: null,
     })
+    mockSetWifiCred.mockReturnValue(<div>Mock SetWifiCred</div>)
   })
 
   it('should render text and buttons', () => {
-    const [{ getByText, getByRole }] = render(
-      '/network-setup/wifi/select-auth-type/mockWiFi'
-    )
-    getByText('Connect to mockWiFi')
+    const [{ getByText, getByRole }] = render(props)
+    getByText('Connect to mockWifi')
     getByRole('button', { name: 'Back' })
     getByRole('button', { name: 'Next' })
     getByText('Select authentication method for your selected network.')
@@ -71,42 +78,29 @@ describe('SelectAuthenticationType', () => {
   })
 
   it('when tapping back button, call a mock function', () => {
-    const [{ getByRole }] = render(
-      '/network-setup/wifi/select-auth-type/mockWiFi'
-    )
+    const [{ getByRole }] = render(props)
+    const button = getByRole('button', { name: 'Back' })
+    fireEvent.click(button)
+    expect(mockPush).toHaveBeenCalledWith('/network-setup/wifi/set-wifi-ssid')
+  })
+
+  it('when tapping back button, call a mock function - fromWifiList', () => {
+    props.fromWifiList = true
+    const [{ getByRole }] = render(props)
     const button = getByRole('button', { name: 'Back' })
     fireEvent.click(button)
     expect(mockPush).toHaveBeenCalledWith('/network-setup/wifi')
   })
 
   it('when tapping next button, call a mock function - wpa', () => {
-    const [{ getByRole }] = render(
-      '/network-setup/wifi/select-auth-type/mockWiFi'
-    )
+    const [{ getByText, getByRole }] = render(props)
     const button = getByRole('button', { name: 'Next' })
     fireEvent.click(button)
-    expect(mockPush).toHaveBeenCalledWith(
-      '/network-setup/wifi/set-wifi-cred/mockWiFi/wpa'
-    )
-  })
-
-  it('when tapping next button, call a mock function - none', () => {
-    const [{ getByRole }] = render(
-      '/network-setup/wifi/select-auth-type/mockWiFi'
-    )
-    const authButton = getByRole('button', { name: 'None' })
-    fireEvent.click(authButton)
-    const button = getByRole('button', { name: 'Next' })
-    fireEvent.click(button)
-    expect(mockPush).toHaveBeenCalledWith(
-      '/network-setup/wifi/set-wifi-cred/mockWiFi/none'
-    )
+    getByText('Mock SetWifiCred')
   })
 
   it('when tapping connect via usb button, call a mock function', () => {
-    const [{ getByRole }] = render(
-      '/network-setup/wifi/select-auth-type/mockWiFi'
-    )
+    const [{ getByRole }] = render(props)
     const button = getByRole('button', { name: 'Connect via USB' })
     fireEvent.click(button)
     expect(mockPush).toHaveBeenCalledWith('/network-setup/usb')
